feat(bh3_valkyries): 优化角色信息展示并添加详情页面
- 修正角色列表信息图片生成逻辑
- 新增单个角色详细信息图片生成功能
- 优化用户指令处理，增加错误提示
- 调整角色搜索和结果显示方式

+++ src/plugins/bh3_valkyries.py
@@ -1,7 +1,7 @@
-import asyncio
 import json
 import uuid
 import random
+import asyncio
 from pathlib import Path
 from nonebot import logger
 from datetime import datetime
@@ -10,19 +10,20 @@ from nonebot.plugin import on_command
 from src.utils.audio import download_audio
 from src.configs.path_config import temp_path
 from src.clover_image.delete_file import delete_file
-from src.bh3_valkyries.base import valkyrie_info_img
+from src.bh3_valkyries.base import valkyries_info_img, valkyrie_info_img
 from src.bh3_valkyries import BH3_User_Assistant, BH3_User_Valkyries
 from nonebot.adapters.qq import   MessageSegment,MessageEvent, Message
-from src.bh3_valkyries.data_base import BH3_Data_base, get_valkyrie_info
+from src.bh3_valkyries.data_base import BH3_Data_base, get_valkyrie_info, get_valkyries_data_ext
 
 bh3_valkyries = on_command("今日助理",aliases={"我的助理"},rule=to_me(),priority=1,block=True)
 @bh3_valkyries.handle()
 async def handle_function(message: MessageEvent):
     cmd = message.get_plaintext().split()
     user_id = message.get_user_id()
-    
-    today = datetime.now().timestamp()
-    current_time = datetime.now().date()
+    # 获取当前时间戳
+    current_time = datetime.now().timestamp()
+    # 获取当前日期
+    today = datetime.now().date()
     
     # 合并用户数据查询
     user_assistant, user_all_valkyries = await asyncio.gather(
@@ -34,7 +35,10 @@ async def handle_function(message: MessageEvent):
         # 获取用户数据
         user_assistant = await BH3_User_Assistant.get_user_data(user_id)
 
-        if len(cmd) == 2 and cmd[1] != "<女武神名称/ID>":
+        if len(cmd) >= 2:
+            if cmd[1] == "<女武神名称/ID>" or len(cmd) > 2:
+                await bh3_valkyries.finish("指令有误 \nTips: 指令：/今日助理 <角色关键字/ID> \n例：/今日助理 979 来设定今日助理哦~!")
+            
             # 设定指定女武神为助理
             user_all_valkyries = await BH3_User_Valkyries.get_user_all_valkyries(user_id)
             if not user_all_valkyries:
@@ -49,12 +53,11 @@ async def handle_function(message: MessageEvent):
             elif isinstance(ids, list):
                 # 搜索结果为多个
                 temp_info_img_path = Path(temp_path) / f"bh3_valkyries_info_{user_id}_{today}_{uuid.uuid4().hex}.png"
-                result = await valkyrie_info_img(
+                result = await valkyries_info_img(
                     await BH3_Data_base.search_user_valkyries_by_id(
                         user_all_valkyries,
                         data, 
-                        ids
-                        ),
+                        ids),
                     temp_info_img_path, "搜索结果")
                 if result:
                      r_msg = Message([
@@ -87,12 +90,12 @@ async def handle_function(message: MessageEvent):
                         assistant_id = ids, 
                         last_set_time = current_time, 
                         first_set_time = current_time
-                        )
+                    )
                     await BH3_User_Valkyries.update_user_valkyrie_data(
                         user_id, 
                         valkyrie_id = ids, 
                         favorability = ids_valkyries.favorability + 1
-                        )
+                    )
                     logger.debug(f"已设置女武神ID:{ids}，好感度+1")
                     await bh3_valkyries.finish(f"已设置女武神ID:{ids}，好感度+1")
                 else:
@@ -107,7 +110,7 @@ async def handle_function(message: MessageEvent):
                             user_id,
                             valkyrie_id = ids,
                             favorability = ids_valkyries.favorability - 1 if ids_valkyries.favorability > 0 else 0
-                            )
+                        )
                         logger.debug(f"已设置女武神ID:{ids}\n当前连续设置次数:{days_diff}")
                         await bh3_valkyries.finish(f"已设置女武神ID:{ids}\n当前连续设置次数:{days_diff} 好感度减少 \n女武神也需要休息哦~！")
                     else:
@@ -115,11 +118,9 @@ async def handle_function(message: MessageEvent):
                             user_id,
                             valkyrie_id = ids,
                             favorability = ids_valkyries.favorability + days_diff
-                            )
+                        )
                         logger.debug(f"用户{user_id} 当前连续设置次数:{days_diff}好感+{days_diff}")
                         await bh3_valkyries.finish(f"已设置女武神ID:{ids}，\n当前连续设置次数:{days_diff}好感+{days_diff}")
-        else:
-            bh3_valkyries.finish("指令有误 \nTips: 指令：/今日助理 <角色关键字/ID> \n例：/今日助理 979 来设定今日助理哦~!")
         
         # 检查是否已获取过助理且最后获取时间是今天
         if user_assistant:
@@ -188,24 +189,50 @@ async def handle_function(message: MessageEvent):
 
             r_data = random.choice(audio_links)
             output_silk_path = await download_audio(r_data['url'])
-            r_msg = Message([
-                 MessageSegment.image(img_link),
-                 MessageSegment.text(r_data['text'])
-            ])
-            await bh3_valkyries.send(r_msg)
 
-            if output_silk_path is None:
-                logger.error("下载音频失败")
-                await bh3_valkyries.finish("下载音频失败，请稍后再试...")
-            await bh3_valkyries.send(MessageSegment.file_audio(Path(output_silk_path)))
-            await delete_file(output_silk_path)
-            await bh3_valkyries.finish()
+            ext_str = valkyrie_info['ext']
+            ext_data = json.loads(ext_str)
+            text_array = json.loads(ext_data['c_18']['filter']['text'])
+            summary = valkyrie_info['summary']
+            no_property_list = ["角色/","初始阶级/"]
+            text_array = [item for item in text_array if not any(prop in item for prop in no_property_list)]
+            # 将/替换为：
+            text_array = [item.replace("/", "：") for item in text_array]
+
+            img_data = {
+                "img_link": img_link,
+                "title": valkyrie_info['title'],
+                "name": await get_valkyries_data_ext(ext_str, "角色/"),
+                "text_array": text_array,
+                "summary": summary,
+                "text": r_data['text']
+            }
+            valkyrie_info_img_path = Path(temp_path) / f"bh3_valkyrie_info_{user_id}_{today}_{uuid.uuid4().hex}.png"
+            if await valkyrie_info_img(img_data, valkyrie_info_img_path):
+                r_msg = Message([
+                     MessageSegment.file_image(valkyrie_info_img_path),
+                     MessageSegment.text(f"\n今日获得女武神！{summary}")
+                ])
+                await bh3_valkyries.send(r_msg)
+
+                if output_silk_path is None:
+                    logger.error("下载音频失败")
+                    await bh3_valkyries.finish("下载音频失败，请稍后再试...")
+                await bh3_valkyries.send(MessageSegment.file_audio(Path(output_silk_path)))
+                await delete_file(valkyrie_info_img_path)
+                await delete_file(output_silk_path)
+                await bh3_valkyries.finish()
+            else:
+                await bh3_valkyries.finish("女武神信息生成失败，请稍后再试...")
 
         else:
             logger.error("数据格式不正确或为空")
             await bh3_valkyries.finish("错误：获取今日助理数据格式不正确/为空，请稍后再试...")
 
     elif cmd[0] == "/我的助理":
+        if cmd[1] == "<女武神名称/ID>" or len(cmd) > 2:
+                await bh3_valkyries.finish("指令有误 \nTips: 指令：/我的助理 <角色关键字/ID> \n例：/我的助理 979 来获取助理信息哦~!")
+        
         user_all_valkyries = await BH3_User_Valkyries.get_user_all_valkyries(user_id)
         if not user_all_valkyries:
             await bh3_valkyries.finish("您还没有任何女武神数据，请先获取今日助理！")
@@ -223,12 +250,11 @@ async def handle_function(message: MessageEvent):
             elif isinstance(ids, list):
                 # 搜索结果为多个
                 temp_info_img_path = Path(temp_path) / f"bh3_valkyries_info_{user_id}_{today}_{uuid.uuid4().hex}.png"
-                if await valkyrie_info_img(
+                if await valkyries_info_img(
                     await BH3_Data_base.search_user_valkyries_by_id(
                         user_all_valkyries,
                         data, 
-                        ids
-                        ),
+                        ids),
                     temp_info_img_path,
                     "搜索结果"
                     ):
@@ -257,16 +283,13 @@ async def handle_function(message: MessageEvent):
                         ])
                         await bh3_valkyries.finish(r_msg)
                 await bh3_valkyries.finish("你还没有获取过这个助理哦!~")
-        else:
-            bh3_valkyries.finish("指令有误 \nTips: 指令：/我的助理 <角色关键字/ID> \n例：/我的助理 979 来获取角色数据哦~!")
         
         # 生成助理列表信息图片
         temp_info_img_path = Path(temp_path) / f"bh3_valkyries_info_{user_id}_{today}_{uuid.uuid4().hex}.png"
-        if await valkyrie_info_img(
+        if await valkyries_info_img(
             await BH3_Data_base.search_user_valkyries_by_id(
                 user_all_valkyries, 
-                data
-                ),
+                data),
             temp_info_img_path, 
             "女武神图鉴"
             ):
